feat(danmaku): 集成bangumi用户信息并更新弹幕存储与请求头
- 弹幕实体类中bgmId字段改名为bgmUserId
- BangumiUserinfoVO中id字段类型调整为Integer
- 配置文件application.yaml添加bangumi User Agent配置
- ForumCommentController中设置userId时转换为Long类型
- DanmakuController新增BangumiClient依赖，添加accessToken参数并获取bangumi用户信息
- DanmakuDto去除无用bgmId字段
- DanmakuService方法签名及实现改为接收并存储bgmUserId
- BangumiClientImpl构造函数注入User Agent并设置默认请求头User-Agent为bangumiUserAgent值

+++ flow-client/src/main/java/com/ligg/flowclient/controller/forum/ForumCommentController.java
@@ -47,7 +47,7 @@ public class ForumCommentController {
         ForumCommentEntity forumCommentEntity = new ForumCommentEntity();
         BeanUtils.copyProperties(forumCommentDto, forumCommentEntity);
         forumCommentEntity.setCreateTime(LocalDateTime.now());
-        forumCommentEntity.setUserId(me.id());
+        forumCommentEntity.setUserId(me.id().longValue());
         forumCommentEntity.setNickname(me.nickname());
         forumCommentEntity.setAvatarUrl(me.avatar().large());
         forumCommentEntity.setContent(forumCommentDto.getContent());
